fixed variables
