+++ store/actions/crypto.actions.tsx
@@ -1,6 +1,6 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { Dispatch } from "react";
-import { Action, Cripto } from "../../interfaces/types";
+import { Action } from "../../interfaces/types";
 
 export const READ_DATA = 'READ_DATA'
