+++ store/index.tsx
@@ -1,12 +1,12 @@
 import thunk from "redux-thunk";
 import { applyMiddleware, combineReducers, createStore } from "redux";
-import { CriptoReducer } from "./reducers/cripto.reducers";
+import { CryptoReducer } from "./reducers/crypto.reducers";
 
 export type RootState = ReturnType<typeof RootReducer>
 
 
 const RootReducer = combineReducers({
-    cripto: CriptoReducer,
+    cripto: CryptoReducer,
 });
 
 export default createStore(RootReducer, applyMiddleware(thunk))
